// src/api/v2/suppliers/suppliers.types.ts
export type TSupplierResponse = {
  id: number;
  name: string;
  code: string;
  created_at: string;
  updated_at: string;
  phone?: string | null;
  address?: string | null;
  orders_count?: number | null;
  refund_orders_count?: number | null;
  total_order_amount?: string | number | null;
  total_payment?: string | number | null;
  total_refund?: string | number | null;
  total_remaining?: string | number | null;
  purchases_count?: number | null;
  returns_count?: number | null;
  net_purchases_count?: number | null;
  total_purchases?: string | number | null;
  total_returns?: string | number | null;
  net_balance?: string | number | null;
  paid?: string | number | null;
  remaining?: string | number | null;
};